requirements.txt
flask
flask_restful
mecab-python3
